schema: ogrp: Change ephemeris parametters units from semi-circles to radians
The issue of ephemeris units does'nt match with ogre units was detected.
We observe at  different systems, units of some of the ephemeris
parameters.
systems:     Units:
ogre         radian
ogrp         semi-circles
GPS_ICD      semi-circles
Rinex        radian
Spirent      semi-circles
Groov's book radian
http://gps.sref.info/course/2d3.html : example out of GPS ephemris from GPS
receivers which uses radian.

if we follow the ogrp schema as semi-circles unit, then we have to consider the similar
change in rtk and in ogre. This is not that usefull to change the unit only to
match the schema. Hence, we change the ogrp schema to radian.

Change-Id: I781c496d5467a8087425cdb5267d8b1ec91b5c42